Dynamic Spline for hose jet
Got a spline that will change dynamically according to the current pressure value and distance to the hit location.

Started work on a render texture for the water to affect and a material for the burnt portion of the wall.

Plan for future is to finish the work on the spline to include a spline mesh to follow it., and if I finish that then I will work on some VFXs to make it come to life