Added Starter Content and more groundwork
Added Starter Content to help progress faster.

Added some groundwork for the hose. I've decided to use a render target and use that to paint the wall.

In terms of score, I could get the amount of grayscale is left, as painted areas would be either white or black. This would also mean that I could dynamically change the material to allow for different areas having different degrees of burns at different play iterations while maintaining the same scoring system.