Updated UV coord generation
Made a function that will find the UV  coordinates around a centre point , using each hit location as a centre. These coordinates are a circular area.

An alternative approach would be save the render target so I can access the pixel location quicker. Then generate a pixel location and read it from that. I'll implement it and see which is more efficient